Support popping multiple controllers in one operation (set backstack)
The NavController issues a navigator.popBackstack() for each and every
destination between the current and desired. This means that router will
step back each Controller in the back stack one by one, invoking their
onCreateView -> onDestroyView.

This is not ideal. Ideally, we want to just "skip" all the controllers
lying in between.

But that requires batching of the popBackstack invocations.
This commit batches the popBackstack calls into a list of transactions
to pop. It uses that list to invoke Router.popToTag(), which does the
batch processing for us. The batching is implemented by posting to an
internal Handler.

Downside of this is that now the popBackStack operation is asynchronous